+++ src/main/java/chess/domain/piece/Blank.java
@@ -1,9 +1,8 @@
 package chess.domain.piece;
 
-import chess.domain.Column;
-import chess.domain.Position;
-import chess.domain.Row;
 import chess.domain.Team;
+import chess.domain.position.Column;
+import chess.domain.position.Position;
 import java.util.EnumMap;
 import java.util.List;
 
@@ -13,17 +12,17 @@ public class Blank extends Piece {
         super(team, position);
     }
 
-    @Override
-    public double getScore() {
-        return 0;
-    }
-
     public static EnumMap<Column, Piece> of(int row, Team team) {
         EnumMap<Column, Piece> blanks = new EnumMap<>(Column.class);
         for (Column column : Column.values()) {
             blanks.put(column, new Blank(team, Position.from(column.getValue() + String.valueOf(row))));
         }
         return blanks;
+    }
+
+    @Override
+    public double getScore() {
+        return 0;
     }
 
     @Override
