+++ src/main/java/chess/domain/piece/MultiStepPiece.java
@@ -1,8 +1,8 @@
 package chess.domain.piece;
 
 import chess.domain.Direction;
-import chess.domain.Position;
 import chess.domain.Team;
+import chess.domain.position.Position;
 import java.util.List;
 
 public abstract class MultiStepPiece extends Piece {
