+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,10 +1,10 @@
 package chess.domain.piece;
 
-import chess.domain.Column;
 import chess.domain.Direction;
-import chess.domain.Position;
-import chess.domain.Row;
 import chess.domain.Team;
+import chess.domain.position.Column;
+import chess.domain.position.Position;
+import chess.domain.position.Row;
 import java.util.EnumMap;
 import java.util.List;
 
@@ -16,17 +16,17 @@ public class Pawn extends MultiStepPiece {
         super(team, position);
     }
 
-    @Override
-    public double getScore() {
-        return PAWN_SCORE;
-    }
-
     public static EnumMap<Column, Piece> of(int row, Team team) {
         EnumMap<Column, Piece> pawns = new EnumMap<>(Column.class);
         for (Column column : Column.values()) {
             pawns.put(column, new Pawn(team, Position.from(column.getValue() + String.valueOf(row))));
         }
         return pawns;
+    }
+
+    @Override
+    public double getScore() {
+        return PAWN_SCORE;
     }
 
     @Override
