+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,7 +1,7 @@
 package chess.domain.piece;
 
-import chess.domain.Position;
 import chess.domain.Team;
+import chess.domain.position.Position;
 import java.util.List;
 
 public abstract class Piece {
